Extract KibanaPluginFeatures type into its own file
