+++ packages/kbn-types/index.ts
@@ -1,9 +1,11 @@
 export {
   KibanaFunctionalPlugin,
   KibanaClassPlugin,
-  KibanaPluginConfig,
-  KibanaPluginFeatures
+  KibanaPluginConfig
 } from '../../platform/server/plugins/types';
+export {
+  KibanaPluginFeatures
+} from '../../platform/server/plugins/KibanaPluginFeatures';
 export { Logger, LoggerFactory } from '../../platform/logging';
 export { Schema, typeOfSchema } from '../../platform/types/schema';
 export { ElasticsearchService } from '../../platform/server/elasticsearch';
